Extract finding TrackerHitResult to helper
